+++ src/components/pages/apps/genes/gsea/bubble/gsea-bubble-panel.tsx
@@ -15,7 +15,7 @@ import { MESSAGE_CHANNEL } from '../../../matcalc/data/data-panel'
 import { ExtScrollCard } from '@/components/ext-scroll-card/ext-scroll-card'
 import { ResizableSidebar } from '@/components/sidebar/resizable-sidebar'
 import { PLOT_ZOOM_CHANNEL } from '../../../matcalc/apps/heatmap/heatmap-panel'
-import { GseaBubblePropsPanel } from '../gsea-plot/bubble/gsea-bubble-props-panel'
+import { GseaBubbleDisplayPropsPanel } from '../gsea-plot/bubble/gsea-bubble-display-props-panel'
 import { GseaBubblePlotSvg } from '../gsea-plot/bubble/gsea-bubble-svg'
 
 export function GseaBubblePanel() {
@@ -84,7 +84,7 @@ export function GseaBubblePanel() {
         <ExtScrollCard className="pb-2">
           <GseaBubblePlotSvg ref={svgRef} />
         </ExtScrollCard>
-        <GseaBubblePropsPanel />
+        <GseaBubbleDisplayPropsPanel />
       </ResizableSidebar>
 
       <FooterPortal className="shrink-0 grow-0 justify-end">